refactor: organizing theme component
