--- src/hooks/useTheme.ts
@@ -1,6 +1,0 @@
-import { useContext } from 'react';
-import { ThemeContext, ThemeContextProps } from '../lib/contexts/ThemeContext';
-
-export function useTheme(): ThemeContextProps {
-  return useContext(ThemeContext);
-}
