+++ src/lib/components/Flowbite/index.tsx
@@ -1,18 +1,17 @@
 import { FC, HTMLAttributes, useLayoutEffect, useMemo } from 'react';
-import useDarkMode from '../../../hooks/useDarkMode';
-import { ThemeContext } from '../../contexts/ThemeContext';
+import { ThemeContext, useThemeMode } from './ThemeContext';
 import { mergeDeep } from '../../helpers/mergeDeep';
 import defaultTheme from '../../theme/default';
 
 interface FlowbiteProps extends HTMLAttributes<HTMLDivElement> {
   children: React.ReactNode;
-  theme?: any;
+  theme?: object;
   dark?: boolean;
   usePreferences?: boolean;
 }
 
 export const Flowbite: FC<FlowbiteProps> = ({ children, theme: customTheme, dark, usePreferences = true }) => {
-  const [mode, setMode, toggleMode] = useDarkMode(usePreferences);
+  const [mode, setMode, toggleMode] = useThemeMode(usePreferences);
 
   const mergedTheme = mergeDeep(defaultTheme, customTheme);
